pyproject.toml
[project]
name = "cdp"
version = "0.1.0"
description = "热点话题感知-过滤-内容生成-分发平台"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Task queue
    "celery[redis]>=5.4",
    # HTTP
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    # AI
    "anthropic>=0.42",
    # Utils
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[project.scripts]
cdp-api = "app.main:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
